Fixed separation of concerns between virtual-table and filterbox
