+++ src/components/filterbox.jsx
@@ -1,47 +1,92 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import Column from 'react-virtualized/dist/es/Table/Column';
 
 import withApp from './with-app';
 import withModel from './with-model';
 import withLayout from './with-layout';
 
 import './filterbox.css';
-import VirtualGrid from './grid';
+import VirtualTable from './virtual-table';
 
 const KEY_ENTER = 13;
 
 function preventDefaultFn(event) {
   event.stopPropagation();
 }
+
+function listboxNameColumnGetter({ rowData }) {
+  return rowData ? rowData[0].qText : '...';
+}
+
+// eslint-disable-next-line react/prop-types
+function nameCellRenderer({ rowData }) {
+  if (!rowData) {
+    return (<div>...</div>);
+  }
+  let title;
+  if (!rowData[0].qNum || rowData[0].qNum === 'NaN') {
+    title = `'${rowData[0].qText ? rowData[0].qText : ''}' (No numerical representation)`;
+  } else {
+    title = `'${rowData[0].qText}' (Numerical representation: ${rowData[0].qNum})`;
+  }
+  return (
+    <div title={title}>
+      {rowData[0].qText ? rowData[0].qText : '< empty >'}
+    </div>
+  );
+}
+
+function listboxFrequencyColumnGetter({ rowData }) {
+  return rowData ? `${rowData[0].qFrequency || '1'}x` : '...';
+}
+
+// eslint-disable-next-line react/prop-types
+function freqCellRenderer({ rowData }) {
+  if (!rowData) {
+    return (<div>...</div>);
+  }
+  if (rowData[0].qFreq > 1) {
+    return (
+      <div title={`This value occurs ${rowData[0].qFreq} times`}>
+        {`${rowData[0].qFreq}x`}
+      </div>
+    );
+  }
+  return <div title="This value occurs only once"> 1x </div>;
+}
+
+function rowRenderer({
+  defaultProps, rowData, style, columns, className, key,
+}) {
+  const classes = `item state-${rowData ? rowData[0].qState : 'Loading'} ${className}`;
+
+  return (
+    <div
+      {...defaultProps}
+      className={classes}
+      key={key}
+      role="row"
+      style={style}
+    >
+      {columns}
+    </div>
+  );
+}
+
+function noRowsRenderer() {
+  return (<div className="no-values">No values</div>);
+}
+
+
 export class Filterbox extends React.Component {
   constructor() {
     super();
-    this.state = {};
+    this.onRowClick = this.onRowClick.bind(this);
+    this.onSearchClick = this.onSearchClick.bind(this);
   }
 
-  componentWillMount() {
-    const { alignTo } = this.props;
-    if (alignTo) {
-      const rect = alignTo.getBoundingClientRect();
-      this.setState({
-        style: {
-          position: 'absolute',
-          top: rect.top + rect.height + window.scrollY,
-          left: rect.left + window.scrollX - 1,
-          width: rect.width + 2, // 1px border x2
-        },
-      });
-    }
-  }
-
-  componentWillUnmount() {
-    const { onClose } = this.props;
-    if (onClose) {
-      onClose();
-    }
-  }
-
-  search(evt) {
+  onSearchClick(evt) {
     const { value } = evt.target;
     const { keyCode } = evt;
     const { model } = this.props;
@@ -57,61 +102,66 @@ export class Filterbox extends React.Component {
     }
   }
 
-  toggleValue(event, item) {
-    event.stopPropagation();
+  onRowClick({ rowData }) {
     const { model } = this.props;
-    model.selectListObjectValues('/qListObjectDef', [item.qElemNumber], true);
+    if (rowData) {
+      if (rowData[0].qState !== 'S') {
+        // eslint-disable-next-line no-param-reassign
+        rowData[0].qState = 'S'; // For fast visual feedback, this will be overwritten when the new layout comes.
+      } else {
+        // eslint-disable-next-line no-param-reassign
+        rowData[0].qState = 'O'; // For fast visual feedback, this will be overwritten when the new layout comes.
+      }
+      this.forceUpdate(); // Force a rerender to ge the previous changes to apear
+
+      model.selectListObjectValues('/qListObjectDef', [rowData[0].qElemNumber], true);
+    }
   }
 
   render() {
-    const { style } = this.state;
     const { layout, model } = this.props;
     if (!layout) {
       return null;
     }
-    console.log(layout);
+
     if (!layout.qListObject.qDataPages || layout.qListObject.qDataPages.length === 0) {
       return null;
     }
-    const items = layout.qListObject.qDataPages[0].qMatrix.map((matrixItem) => {
-      const item = matrixItem[0];
-      const classes = `item state-${item.qState}`;
-      return (
-        <li
-          key={item.qElemNumber}
-          className={classes}
-          onClick={event => this.toggleValue(event, item)}
-        >
-          {item.qText
-            + (item.qFrequency && item.qFrequency !== '1'
-              ? ` (${item.qFrequency}x)`
-              : '')}
-        </li>
-      );
-    });
-    const search = (
-      <input
-        autoFocus
-        onKeyUp={evt => this.search(evt)}
-        className="search"
-        placeholder="Search (wildcard)"
-      />
-    );
-    if (!items.length) {
-      return (
-        <div role="Listbox" tabIndex="-1" className="filterbox empty" style={style} onClick={preventDefaultFn}>
-          {search}
-          <ul className="items">
-            <li>No values.</li>
-          </ul>
-        </div>
-      );
-    }
+
     return (
-      <div role="Listbox" tabIndex="-1" className="filterbox" style={style} onClick={preventDefaultFn}>
-        {search}
-        <div className="virtualgrid">
-          <VirtualGrid layout={layout} model={model} />
+      <div role="Listbox" tabIndex="-1" className="filterbox" onClick={preventDefaultFn}>
+        <input
+          onKeyUp={this.onSearchClick}
+          className="search"
+          placeholder="Search (wildcard)"
+        />
+        <div className="virtualtable">
+          <VirtualTable
+            layout={layout}
+            model={model}
+            onRowClick={this.onRowClick}
+            rowRenderer={rowRenderer}
+            noRowsRenderer={noRowsRenderer}
+          >
+            <Column
+              label="Name"
+              dataKey="name"
+              width={100}
+              flexGrow={1}
+              flexShrink={1}
+              cellDataGetter={listboxNameColumnGetter}
+              cellRenderer={nameCellRenderer}
+            />
+            <Column
+              width={40}
+              label="Description"
+              dataKey="description"
+              flexGrow={1}
+              flexShrink={0}
+              cellDataGetter={listboxFrequencyColumnGetter}
+              cellRenderer={freqCellRenderer}
+            />
+          </VirtualTable>
         </div>
       </div>
     );
@@ -121,16 +171,12 @@ export class Filterbox extends React.Component {
 Filterbox.defaultProps = {
   model: null,
   layout: null,
-  alignTo: null,
-  onClose: null,
 };
 
 Filterbox.propTypes = {
   model: PropTypes.object,
   layout: PropTypes.object,
   field: PropTypes.string.isRequired, // eslint-disable-line react/no-unused-prop-types
-  alignTo: PropTypes.object,
-  onClose: PropTypes.func,
 };
 
 export default withApp(withModel(withLayout(Filterbox), (app, props) => app.getOrCreateListbox(props.field)));
